+++ lib/user.js
@@ -48,10 +48,9 @@ class User {
     return Promise.resolve(new User())
   }
 
-  getServiceProviders() {
-    const token = localStorage.getItem('token-fc')
+  getServiceProviders(token) {
     if (!token) {
-      return Promise.reject()
+      return Promise.reject(new Error('Token not provided'))
     }
     return axios.get('https://partenaires.dev.dev-franceconnect.fr/oauth/v1/userinfo', {
       headers: {Authorization: 'Bearer ' + token}
