+++ lib/utils.js
@@ -11,6 +11,23 @@ const Utils = {
     }
     return ''
   },
+
+  getQueryVariable(variable) {
+    if (typeof window === 'undefined') {
+      return null
+    }
+
+    const query = window.location.search.substring(1)
+    const vars = query.split('&')
+
+    for (let i = 0; i < vars.length; i++) {
+      const pair = vars[i].split('=')
+      if (decodeURIComponent(pair[0]) === variable) {
+        return decodeURIComponent(pair[1])
+      }
+    }
+  },
+
   // TODO remove and replace by lodash utilities
   deepSetInState(name, value, stateCopy) {
     name.split('.').reduce((acc, element) => {
