+++ package.json
@@ -28,7 +28,6 @@
     "jsdom": "^11.6.2",
     "lodash": "^4.17.10",
     "next": "^5.1.0",
-    "lodash": "^4.17.10",
     "nock": "^9.2.3",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
